[CV-5702]: Production deployment hung during stackdriver log verification step(s) (#17687)
* [CV-5702]: Production deployment hung during stackdriver log verification step(s)

* [CV-5702]: fix test

* [CV-5702]: unit test

+++ 270-verification/src/main/java/io/harness/jobs/workflow/collection/WorkflowDataCollectionJob.java
@@ -1,21 +1,36 @@
 package io.harness.jobs.workflow.collection;
 
-import com.google.inject.Inject;
+import static software.wings.common.VerificationConstants.DELAY_MINUTES;
+import static software.wings.common.VerificationConstants.WORKFLOW_CV_COLLECTION_CRON_GROUP;
 
+import com.google.inject.Inject;
+import com.google.inject.name.Named;
+
+import io.harness.managerclient.VerificationManagerClient;
+import io.harness.managerclient.VerificationManagerClientHelper;
+import io.harness.mongo.iterator.MongoPersistenceIterator.Handler;
+import io.harness.scheduler.PersistentScheduler;
 import io.harness.serializer.JsonUtils;
 import io.harness.service.intfc.ContinuousVerificationService;
 import lombok.extern.slf4j.Slf4j;
 import org.quartz.Job;
 import org.quartz.JobExecutionContext;
 import org.quartz.SchedulerException;
+import software.wings.beans.FeatureName;
 import software.wings.service.impl.analysis.AnalysisContext;
+
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
 
 /**
  * Created by Pranjal on 02/06/2019
  */
 @Slf4j
-public class WorkflowDataCollectionJob implements Job {
+public class WorkflowDataCollectionJob implements Job, Handler<AnalysisContext> {
+  @Inject private VerificationManagerClientHelper verificationManagerClientHelper;
+  @Inject private VerificationManagerClient verificationManagerClient;
   @Inject private ContinuousVerificationService continuousVerificationService;
+  @Inject @Named("BackgroundJobScheduler") private PersistentScheduler jobScheduler;
 
   @Override
   public void execute(JobExecutionContext jobExecutionContext) {
@@ -23,20 +38,53 @@ public class WorkflowDataCollectionJob implements Job {
     try {
       String params = jobExecutionContext.getMergedJobDataMap().getString("jobParams");
       AnalysisContext context = JsonUtils.asObject(params, AnalysisContext.class);
-      logger.info("Executing Workflow data collection job with params : {} and context : {}", params, context);
-      boolean jobTriggered = continuousVerificationService.triggerWorkflowDataCollection(context);
-      if (!jobTriggered) {
-        deleteJob(jobExecutionContext);
+      if (!verificationManagerClientHelper
+               .callManagerWithRetry(verificationManagerClient.isFeatureEnabled(
+                   FeatureName.WORKFLOW_DATA_COLLECTION_ITERATOR, context.getAccountId()))
+               .getResource()) {
+        logger.info("Executing Workflow data collection job with params : {} and context : {}", params, context);
+        boolean jobTriggered = continuousVerificationService.triggerWorkflowDataCollection(context);
+        if (!jobTriggered) {
+          deleteJob(jobExecutionContext);
+        }
+        logger.info("Triggering scheduled job with params {}", params);
+      } else {
+        logger.info("for {} the iterator will handle data collection", context.getStateExecutionId());
       }
-      logger.info("Triggering scheduled job with params {}", params);
     } catch (Exception ex) {
       logger.error("Data Collection cron failed with error", ex);
     }
   }
 
+  @Override
+  public void handle(AnalysisContext context) {
+    if (!verificationManagerClientHelper
+             .callManagerWithRetry(verificationManagerClient.isFeatureEnabled(
+                 FeatureName.WORKFLOW_DATA_COLLECTION_ITERATOR, context.getAccountId()))
+             .getResource()) {
+      logger.info("for {} the cron will handle data collection", context.getStateExecutionId());
+      return;
+    }
+    if (Instant.now().isBefore(Instant.ofEpochMilli(context.getCreatedAt()).plus(DELAY_MINUTES, ChronoUnit.MINUTES))) {
+      logger.info("for {} the delay of {} mins hasn't reached yet", context.getStateExecutionId(), DELAY_MINUTES);
+      return;
+    }
+
+    logger.info("Executing Workflow data collection iterator for context : {}", context);
+    boolean jobTriggered = continuousVerificationService.triggerWorkflowDataCollection(context);
+    if (!jobTriggered) {
+      continuousVerificationService.markWorkflowDataCollectionDone(context);
+      jobScheduler.deleteJob(context.getStateExecutionId(),
+          context.getStateType().name().toUpperCase() + WORKFLOW_CV_COLLECTION_CRON_GROUP);
+    }
+  }
+
   private void deleteJob(JobExecutionContext jobExecutionContext) {
     try {
+      String params = jobExecutionContext.getMergedJobDataMap().getString("jobParams");
+      AnalysisContext context = JsonUtils.asObject(params, AnalysisContext.class);
       jobExecutionContext.getScheduler().deleteJob(jobExecutionContext.getJobDetail().getKey());
+      continuousVerificationService.markWorkflowDataCollectionDone(context);
       logger.info("Deleting Data Collection job for context {}", jobExecutionContext);
     } catch (SchedulerException e) {
       logger.error("Unable to clean up cron", e);
